Fix Travis test failures
Attempts to fix #171

These set of changes attempt to factor our most of the functionality of
ResourceCompiler into a library so that it's facilitates easier testing. That
way, the test driver program invokes a function call instead of using the more
error-prone std::system. This is WIP and there are quite a few test failures.