Support multiple listeners
Proof of concept implementation of supporting multiple listeners. The multiple
listeners implementation was present but didn't work as evidenced in bugs #83
and #95.

This adds a templated AddFrameworkListener that takes callables with distinct
addresses - free functions and functors. To achieve this, we make use of the
compile time traits utility FunctionTraits.h

Also adds implementation of the FrameworkToken object and an internal atomic id
for listeners.

The FrameworkListenerTest is changed to support just the multiple listeners for
the time being.